Tweaked semantics of when sizes are returned.
Now, if a view is not attached to a window, it will assume that it
should be sized according to the main window. Sizes of 0,0 will only be
returned if we know that you’re on a secondary display, and if you try
to get a transform for a view that’s on a secondary display, it will
just return the view’s current transform.